+++ src/main/kotlin/io/github/malczuuu/nullmarked/internal/PackageInfoCommons.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.github.malczuuu.nullmarked
+package io.github.malczuuu.nullmarked.internal
 
 import java.io.File